// back/templates/front/src/app/api/models/field.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { FieldType } from '../models/field-type';
import { Option } from '../models/option';
export interface Field {
  comment?: string | null;
  defaultValue?: string | null;
  description?: string | null;
  id: string | null;
  isDeleted: boolean;
  isReadOnly: boolean;
  isRequired: boolean;
  label: string | null;
  options?: Array<Option> | null;
  order: number;
  photo?: string | null;
  type: FieldType;
}